Add a timeout to SERCOM::readDataWIRE to narrow down I2C hangs.

+++ packages/adafruit/hardware/samd/1.0.22/cores/arduino/SERCOM.h
@@ -205,7 +205,7 @@ class SERCOM
 		bool isMasterReadOperationWIRE( void ) ;
     bool isRXNackReceivedWIRE( void ) ;
 		int availableWIRE( void ) ;
-		uint8_t readDataWIRE( void ) ;
+		int16_t readDataWIRE( void ) ;
 
 	private:
 		Sercom* sercom;
